Validate hypertoroidal Dirac conversion counts (#2389)

+++ tests/distributions/test_hypertoroidal_dirac_distribution.py
@@ -65,16 +65,18 @@ class TestHypertoroidalDiracDistribution(TestAbstractDiracDistribution):
         npt.assert_allclose(dist.d, array([0.1, 0.2, 0.3]))
 
     def test_from_distribution_sampling_1d(self):
-        n_particles = 5
+        n_particles = np.int64(5)
         vm = VonMisesDistribution(array(0.2), array(1.5))
 
         wd = HypertoroidalDiracDistribution.from_distribution(vm, n_particles)
 
         self.assertIsInstance(wd, HypertoroidalDiracDistribution)
         self.assertEqual(wd.dim, 1)
-        self.assertEqual(wd.d.shape, (n_particles,))
-        self.assertEqual(wd.w.shape, (n_particles,))
-        npt.assert_array_almost_equal(wd.w, array([1.0 / n_particles] * n_particles))
+        self.assertEqual(wd.d.shape, (int(n_particles),))
+        self.assertEqual(wd.w.shape, (int(n_particles),))
+        npt.assert_array_almost_equal(
+            wd.w, array([1.0 / n_particles] * int(n_particles))
+        )
         npt.assert_array_almost_equal(wd.d, mod(wd.d, 2.0 * pi))
 
     def test_from_distribution_validates_sampling_particle_count(self):
